+++ resources/lib/UnityDeployment.php
@@ -42,89 +42,118 @@ class UnityDeployment
     }
 
     /** @param mixed[] $x */
-    private static function doesArrayHaveOnlyIntegerValues(array $x): bool
+    private static function assertAllValuesAreInts(array $x, string $name): void
     {
         foreach ($x as $value) {
-            if (!is_int($value)) {
-                return false;
-            }
+            assert(is_int($value), "all values of $name must be integers");
         }
-        return true;
     }
 
-    /** @param int[] $x */
-    private static function isArrayMonotonicallyIncreasing(array $x): bool
-    {
-        if (count($x) <= 1) {
-            return true;
+    /** @param mixed[] $x */
+    private static function assertArrayIsOneOrMoreMonotonicallyIncreasingIntegers(
+        array $x,
+        string $name,
+    ): void {
+        self::assertArrayNotEmpty($x, $name);
+        self::assertAllValuesAreInts($x, $name);
+        if (count($x) === 1) {
+            return;
         }
         $remaining_values = $x;
         $last_value = array_shift($remaining_values);
         while (count($remaining_values)) {
             $this_value = array_shift($remaining_values);
-            if ($this_value < $last_value) {
-                return false;
-            }
+            assert($this_value >= $last_value, "$name must be monotonically increasing");
             $last_value = $this_value;
         }
-        return true;
     }
 
     /** @param mixed[] $CONFIG */
     private static function validateConfig(array $CONFIG): void
     {
+        if (ini_get("zend.assertions") !== "1") {
+            throw new InvalidConfigurationException("zend.assertions must be set to 1");
+        }
+        if (ini_get("assert.exception") !== "1") {
+            throw new InvalidConfigurationException("assert.exception must be set to 1");
+        }
+        try {
+            self::validateExpiryConfig($CONFIG);
+            self::validateSmtpConfig($CONFIG);
+        } catch (\Throwable $e) {
+            throw new InvalidConfigurationException($e->getMessage());
+        }
+    }
+
+    /** @param mixed[] $CONFIG */
+    private static function validateExpiryConfig(array $CONFIG): void
+    {
         $idlelock_warning_days = $CONFIG["expiry"]["idlelock_warning_days"];
         $idlelock_day = $CONFIG["expiry"]["idlelock_day"];
         $disable_warning_days = $CONFIG["expiry"]["disable_warning_days"];
         $disable_day = $CONFIG["expiry"]["disable_day"];
-        if (count($idlelock_warning_days) === 0) {
-            throw new InvalidConfigurationException(
-                '$CONFIG["expiry"]["idlelock_warning_days"] must not be empty!',
-            );
-        }
-        if (count($disable_warning_days) === 0) {
-            throw new InvalidConfigurationException(
-                '$CONFIG["expiry"]["disable_warning_days"] must not be empty!',
-            );
-        }
-        if (!self::doesArrayHaveOnlyIntegerValues($idlelock_warning_days)) {
-            throw new InvalidConfigurationException(
-                '$CONFIG["expiry"]["idlelock_warning_days"] must be a list of integers!',
-            );
-        }
-        if (!self::doesArrayHaveOnlyIntegerValues($disable_warning_days)) {
-            throw new InvalidConfigurationException(
-                '$CONFIG["expiry"]["disable_warning_days"] must be a list of integers!',
-            );
-        }
-        if (!self::isArrayMonotonicallyIncreasing($idlelock_warning_days)) {
-            throw new InvalidConfigurationException(
-                '$CONFIG["expiry"]["idlelock_warning_days"] must be monotonically increasing!',
-            );
-        }
-        if (!self::isArrayMonotonicallyIncreasing($disable_warning_days)) {
-            throw new InvalidConfigurationException(
-                '$CONFIG["expiry"]["disable_warning_days"] must be monotonically increasing!',
-            );
-        }
-
+        self::assertArrayIsOneOrMoreMonotonicallyIncreasingIntegers(
+            $idlelock_warning_days,
+            '$CONFIG["expiry"]["idlelock_warning_days"]',
+        );
+        self::assertArrayIsOneOrMoreMonotonicallyIncreasingIntegers(
+            $disable_warning_days,
+            '$CONFIG["expiry"]["disable_warning_days"]',
+        );
         $final_disable_warning_day = _array_last($disable_warning_days);
         $final_idlelock_warning_day = _array_last($idlelock_warning_days);
-        if ($disable_day <= $final_disable_warning_day) {
-            throw new InvalidConfigurationException(
-                "disable day must be greater than the last disable warning day",
-            );
-        }
-        if ($idlelock_day <= $final_idlelock_warning_day) {
-            throw new InvalidConfigurationException(
-                "idlelock day must be greater than the last idlelock warning day",
-            );
-        }
-        if ($disable_day <= $idlelock_day) {
-            throw new InvalidConfigurationException(
-                "disable day must be greater than idlelock day",
-            );
-        }
+        assert(
+            $disable_day > $final_disable_warning_day,
+            "disable day must be greater than the last disable warning day",
+        );
+        assert(
+            $idlelock_day > $final_idlelock_warning_day,
+            "idlelock day must be greater than the last idlelock warning day",
+        );
+        assert($disable_day > $idlelock_day, "disable day must be greater than idlelock day");
+    }
+
+    /** @param mixed[] $CONFIG */
+    private static function validateSmtpConfig(array $CONFIG): void
+    {
+        self::assertStringNotEmpty($CONFIG["smtp"]["host"], '$CONFIG["smtp"]["host"]');
+        self::assertIsInt($CONFIG["smtp"]["port"], '$CONFIG["smtp"]["port"]');
+        self::assertOneOf($CONFIG["smtp"]["security"], '$CONFIG["smtp"]["security"]', [
+            "",
+            "tls",
+            "ssl",
+        ]);
+        self::assertIsBool($CONFIG["smtp"]["ssl_verify"], '$CONFIG["smtp"]["ssl_verify"]');
+    }
+
+    /** @param mixed[] $x */
+    private static function assertArrayNotEmpty(array $x, string $name): void
+    {
+        assert(count($x) > 0, "$name must not be empty");
+    }
+
+    private static function assertStringNotEmpty(string $x, string $name): void
+    {
+        assert(!empty($x), "$name must not be empty");
+    }
+
+    private static function assertIsBool(mixed $x, string $name): void
+    {
+        assert(is_bool($x), "$name must be a boolean");
+    }
+
+    private static function assertIsInt(mixed $x, string $name): void
+    {
+        assert(is_int($x), "$name must be a boolean");
+    }
+
+    /** @param mixed[] $options */
+    private static function assertOneOf(string $x, string $name, array $options): void
+    {
+        assert(
+            in_array($x, $options, true),
+            sprintf("%s must be one of %s", $name, _json_encode($options)),
+        );
     }
 
     private static function assertHttpHostValid(string $host): void